svg/animations/end-event.svg

<svg xmlns="http://www.w3.org/2000/svg"><title>Animation triggers on &apos;endEvent&apos;</title><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/><script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/><rect width="0" height="100" fill="green"><animate id="anim" attributeName="visibility" begin="0s" end="2s" to="visible"/><set attributeName="width" begin="anim.endEvent" to="100"/></rect><script>async_test(t =&gt; { document.documentElement.setCurrentTime(1.99); let set = document.querySelector(&apos;set&apos;); set.addEventListener(&apos;beginEvent&apos;, t.step_func_done(function() { assert_equals(set.parentNode.getBBox().width, 100); })); });</script></svg>
